Moving BadgeList and Bookmarks
