+++ amundsen_application/static/js/components/Bookmark/BookmarkIcon/styles.scss
