+++ amundsen_application/static/js/components/Bookmark/MyBookmarks/styles.scss
