+++ amundsen_application/static/js/components/common/ResourceListItem/DashboardListItem/index.spec.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 
 import { Link } from 'react-router-dom';
-import BookmarkIcon from 'components/common/Bookmark/BookmarkIcon';
+import BookmarkIcon from 'components/Bookmark/BookmarkIcon';
 import { ResourceType } from 'interfaces';
 
 import * as ConfigUtils from 'config/config-utils';
