+++ amundsen_application/static/js/features/BadgeList/index.tsx
@@ -10,7 +10,7 @@ import { ResourceType, Badge } from 'interfaces';
 import { updateSearchState } from 'ducks/search/reducer';
 import { UpdateSearchStateRequest } from 'ducks/search/types';
 
-import BadgeList from 'components/common/BadgeList';
+import BadgeList from 'components/BadgeList';
 
 export interface DispatchFromProps {
   onBadgeClick: (badgeText: string) => UpdateSearchStateRequest;
