+++ amundsen_application/static/js/pages/HomePage/index.spec.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 
 import Breadcrumb from 'components/common/Breadcrumb';
-import MyBookmarks from 'components/common/Bookmark/MyBookmarks';
+import MyBookmarks from 'components/Bookmark/MyBookmarks';
 import PopularTables from 'components/common/PopularTables';
 import SearchBar from 'components/common/SearchBar';
 import TagsListContainer from 'components/common/Tags';
